Fix insufficient permissions error on NetworkManager
Currently network manager was complaining that somehow it lacked the
permissions to create new networks. This polkit rule fixes the problem,
the important part here is to ensure that the user goup's matches the
group specified in the polkit rule, it can be any group as long as they
match. Upon applying the rule, there's no need to restart anything,
polkit will reload itself